requirements.txt
pandas==2.2.1
requests==2.31.0
python-telegram-bot==13.15
